[JT-52] style: package 명 변경
- PaymentInfoDto -> PaymentInfoReq
