+++ module-domain/src/main/java/com/devtoon/jtoon/paymentinfo/entity/PG.java
@@ -1,4 +1,4 @@
-package com.devtoon.jtoon.payment.entity;
+package com.devtoon.jtoon.paymentinfo.entity;
 
 import java.util.Arrays;
 import java.util.Collections;
